* bug fix: webapp-config-1.50.16 errors out during uninstall htdocs is a symlink (Gentoo bug #181938)
svn path=/trunk/webapp-config/; revision=56

+++ WebappConfig/server.py
@@ -161,7 +161,7 @@ class Basic:
 
         # is the installation directory empty?
 
-        if not os.listdir(self.__destd):
+        if not os.listdir(self.__destd) and os.path.isdir(self.__destd):
             if not self.__p:
                 os.rmdir(self.__destd)
         else:
